move create_file, update_yaml to utils
