+++ .github/scripts/utils.py
@@ -1,5 +1,6 @@
 import os
 import re
+import yaml
 import json
 import shutil
 import requests
@@ -9,8 +10,7 @@ from io import BytesIO
 import urllib.parse
 from pathlib import Path
 from typing import Dict, Tuple
-
-
+from config import *
 
 def process_friendly_url(friendly_url, replace = "-"):
     # Удаление специальных символов
@@ -343,6 +343,7 @@ def setup_directories(thumbs_dir: str, cars_dir: str) -> None:
         shutil.rmtree(cars_dir)
     os.makedirs(cars_dir)
 
+
 def should_remove_car(car: ET.Element, mark_ids: list, folder_ids: list) -> bool:
     """
     Проверяет, нужно ли удалить машину по заданным критериям.
@@ -366,3 +367,247 @@ def should_remove_car(car: ET.Element, mark_ids: list, folder_ids: list) -> bool
             return True
     
     return False
+
+
+def create_file(car, filename, friendly_url, current_thumbs, existing_files, elements_to_localize, config):
+    vin = car.find('vin').text
+    vin_hidden = process_vin_hidden(vin)
+    # Преобразование цвета
+    color = car.find('color').text.strip().capitalize()
+    model = car.find('folder_id').text.strip()
+    brand = car.find('mark_id').text.strip()
+
+    folder = get_folder(brand, model)
+    color_image = get_color_filename(brand, model, color)
+    if folder and color_image:
+        thumb = f"/img/models/{folder}/colors/{color_image}"
+    else:
+        print("")
+        errorText = f"VIN: {vin}. Не хватает модели: {model} или цвета: {color}"
+        print(errorText)
+        print("")
+        with open('output.txt', 'a') as file:
+            file.write(f"{errorText}\n")
+        # Если 'model' или 'color' не найдены, используем путь к изображению ошибки 404
+        thumb = "/img/404.jpg"
+
+
+    # Forming the YAML frontmatter
+    content = "---\n"
+    # content += "layout: car-page\n"
+    total_element = car.find('total')
+    if total_element is not None:
+        content += f"total: {int(total_element.text)}\n"
+    else:
+        content += "total: 1\n"
+    # content += f"permalink: {friendly_url}\n"
+    content += f"vin_hidden: {vin_hidden}\n"
+
+    h1 = join_car_data(car, 'mark_id', 'folder_id', 'modification_id')
+    content += f"h1: {h1}\n"
+
+    content += f"breadcrumb: {join_car_data(car, 'mark_id', 'folder_id', 'complectation_name')}\n"
+
+    content += f"title: 'Купить {join_car_data(car, 'mark_id', 'folder_id', 'modification_id')} у официального дилера в {dealer.get('where')}'\n"
+
+    description = (
+        f'Купить автомобиль {join_car_data(car, "mark_id", "folder_id")}'
+        f'{" " + car.find("year").text + " года выпуска" if car.find("year").text else ""}'
+        f'{", комплектация " + car.find("complectation_name").text if car.find("complectation_name").text != None else ""}'
+        f'{", цвет - " + car.find("color").text if car.find("color").text != None else ""}'
+        f'{", двигатель - " + car.find("modification_id").text if car.find("modification_id").text != None else ""}'
+        f' у официального дилера в г. {dealer.get("city")}. Стоимость данного автомобиля {join_car_data(car, "mark_id", "folder_id")} – {car.find("priceWithDiscount").text}'
+    )
+    content += f"description: '{description}'\n"
+
+    description = ""
+
+    for elem_name in elements_to_localize:
+        elem = car.find(elem_name)
+        localize_element_text(elem)
+
+    color = car.find('color').text.strip().capitalize()
+    encountered_tags = set()  # Создаем множество для отслеживания встреченных тегов
+
+    for child in car:
+        # Skip nodes with child nodes (except image_tag) and attributes
+        if list(child) and child.tag != f'{config["image_tag"]}s':
+            continue
+        if child.tag == 'total':
+            continue
+        if child.tag == 'folder_id':
+            content += f"{child.tag}: '{child.text}'\n"
+        elif child.tag == f'{config["image_tag"]}s':
+            images = [img.text for img in child.findall(config['image_tag'])]
+            thumbs_files = createThumbs(images, friendly_url, current_thumbs, config['thumbs_dir'], config['skip_thumbs'])
+            content += f"images: {images}\n"
+            content += f"thumbs: {thumbs_files}\n"
+        elif child.tag == 'color':
+            content += f"{child.tag}: {color}\n"
+            content += f"image: {thumb}\n"
+        elif child.tag == 'extras' and child.text:
+            extras = child.text
+            flat_extras = extras.replace('\n', '<br>\n')
+            content += f"{child.tag}: |\n"
+            for line in flat_extras.split("\n"):
+                content += f"  {line}\n"
+        elif child.tag == config['description_tag'] and child.text:
+            description = child.text
+            flat_description = description.replace('\n', '<br>\n')
+            # content += f"content: |\n"
+            # for line in flat_description.split("\n"):
+                # content += f"  {line}\n"
+        elif child.tag == 'equipment' and child.text:
+            description = child.text
+            flat_description = description.replace('\n', '<br>\n')
+            content += f"{child.tag}: |\n"
+            for line in flat_description.split("\n"):
+                content += f"  {line}\n"
+        else:
+            if child.tag in encountered_tags:  # Проверяем, встречался ли уже такой тег
+                continue  # Если встречался, переходим к следующей итерации цикла
+            encountered_tags.add(child.tag)  # Добавляем встреченный тег в множество
+            if child.text:  # Only add if there's content
+                content += f"{child.tag}: {child.text}\n"
+
+    content += "---\n"
+    content += process_description(description)
+
+    with open(filename, 'w') as f:
+        f.write(content)
+
+    print(filename);
+    existing_files.add(filename)
+
+
+def update_yaml(car, filename, friendly_url, current_thumbs, config):
+
+    with open(filename, "r", encoding="utf-8") as f:
+        content = f.read()
+
+    # Split the content by the YAML delimiter
+    yaml_delimiter = "---\n"
+    parts = content.split(yaml_delimiter)
+
+    # If there's no valid YAML block, raise an exception
+    if len(parts) < 3:
+        raise ValueError("No valid YAML block found in the provided file.")
+
+    # Parse the YAML block
+    yaml_block = parts[1].strip()
+    data = yaml.safe_load(yaml_block)
+
+    total_element = car.find('total')
+    if 'total' in data and total_element is not None:
+        try:
+            car_total_value = int(total_element.text)
+            data_total_value = int(data['total'])
+            data['total'] = data_total_value + car_total_value
+        except ValueError:
+            # В случае, если не удается преобразовать значения в int,
+            # можно оставить текущее значение data['total'] или установить его в 0,
+            # либо выполнить другое действие по вашему выбору
+            pass
+    else:
+        # Если элемент 'total' отсутствует в одном из источников,
+        # можно установить значение по умолчанию для 'total' в data или обработать этот случай иначе
+        data['total'] += 1
+
+    run_element = car.find('run')
+    if 'run' in data and run_element is not None:
+        try:
+            car_run_value = int(run_element.text)
+            data_run_value = int(data['run'])
+            data['run'] = min(data_run_value, car_run_value)
+        except ValueError:
+            # В случае, если не удается преобразовать значения в int,
+            # можно оставить текущее значение data['run'] или установить его в 0,
+            # либо выполнить другое действие по вашему выбору
+            pass
+    else:
+        # Если элемент 'run' отсутствует в одном из источников,
+        # можно установить значение по умолчанию для 'run' в data или обработать этот случай иначе
+        data.setdefault('run', 0)
+
+    priceWithDiscount_element = car.find('priceWithDiscount')
+    if 'priceWithDiscount' in data and priceWithDiscount_element is not None:
+        try:
+            car_priceWithDiscount_value = int(priceWithDiscount_element.text)
+            data_priceWithDiscount_value = int(data['priceWithDiscount'])
+            data['priceWithDiscount'] = min(data_priceWithDiscount_value, car_priceWithDiscount_value)
+            data['sale_price'] = min(data_priceWithDiscount_value, car_priceWithDiscount_value)
+            description = (
+                f'Купить автомобиль {join_car_data(car, "mark_id", "folder_id")}'
+                f'{" " + car.find("year").text + " года выпуска" if car.find("year").text else ""}'
+                f'{", комплектация " + car.find("complectation_name").text if car.find("complectation_name").text != None else ""}'
+                f'{", цвет - " + car.find("color").text if car.find("color").text != None else ""}'
+                f'{", двигатель - " + car.find("modification_id").text if car.find("modification_id").text != None else ""}'
+                f' у официального дилера в г. {dealer.get("city")}. Стоимость данного автомобиля {join_car_data(car, "mark_id", "folder_id")} – {car.find("priceWithDiscount").text}'
+            )
+            data["description"] = description
+        except ValueError:
+            # В случае, если не удается преобразовать значения в int,
+            # можно оставить текущее значение data['priceWithDiscount'] или установить его в 0,
+            # либо выполнить другое действие по вашему выбору
+            pass
+    # else:
+        # Если элемент 'priceWithDiscount' отсутствует в одном из источников,
+        # можно установить значение по умолчанию для 'priceWithDiscount' в data или обработать этот случай иначе
+        # data.setdefault('priceWithDiscount', 0)
+
+    max_discount_element = car.find('max_discount')
+    if 'max_discount' in data and max_discount_element is not None:
+        try:
+            car_max_discount_value = int(max_discount_element.text)
+            data_max_discount_value = int(data['max_discount'])
+            data['max_discount'] = max(data_max_discount_value, car_max_discount_value)
+        except ValueError:
+            # В случае, если не удается преобразовать значения в int,
+            # можно оставить текущее значение data['max_discount'] или установить его в 0,
+            # либо выполнить другое действие по вашему выбору
+            pass
+
+
+    vin = car.find('vin').text
+    vin_hidden = process_vin_hidden(vin)
+    if vin_hidden is not None:
+        # Создаём или добавляем строку в список
+        data['vin_hidden'] += ", " + vin_hidden
+
+    unique_id = car.find('unique_id')
+    if unique_id is not None:
+        if not isinstance(data['unique_id'], str):
+            data['unique_id'] = str(data['unique_id'])
+
+        data['unique_id'] += ", " + str(unique_id.text)
+    else:
+        unique_id = car.find('id')
+        if unique_id is not None:
+            if not isinstance(data['id'], str):
+                data['id'] = str(data['id'])
+
+            data['id'] += ", " + str(unique_id.text)
+
+
+    images_container = car.find(f"{config['image_tag']}s")
+    if images_container is not None:
+        images = [img.text for img in images_container.findall(config['image_tag'])]
+        if len(images) > 0:
+            data.setdefault('images', []).extend(images)
+            # Проверяем, нужно ли добавлять эскизы
+            if 'thumbs' not in data or (len(data['thumbs']) < 5):
+                thumbs_files = createThumbs(images, friendly_url, current_thumbs, config['thumbs_dir'], config['skip_thumbs'])
+                data.setdefault('thumbs', []).extend(thumbs_files)
+
+    # Convert the data back to a YAML string
+    updated_yaml_block = yaml.safe_dump(data, default_flow_style=False, allow_unicode=True)
+
+    # Reassemble the content with the updated YAML block
+    updated_content = yaml_delimiter.join([parts[0], updated_yaml_block, yaml_delimiter.join(parts[2:])])
+
+    # Save the updated content to the output file
+    with open(filename, "w", encoding="utf-8") as f:
+        f.write(updated_content)
+
+    return filename
+
